- Moving the projection calculation of screenCoordinates from int to qreal. This makes our new shiny graticule look better and even seems to result in a tiny little bit better performance -- at least on desktop computers ...
svn path=/trunk/KDE/kdeedu/marble/; revision=958993

+++ src/lib/MarbleMap.cpp
@@ -281,7 +281,7 @@ void MarbleMapPrivate::paintGround( GeoPainter &painter, QRect &dirtyRect )
         drawFog( painter );
 }
 
-void MarbleMapPrivate::paintOverlay( GeoPainter &painter, QRect &dirtyRect)
+void MarbleMapPrivate::paintOverlay( GeoPainter &painter, QRect &dirtyRect )
 {
     if ( !m_viewParams.mapTheme() ) {
         return;
@@ -869,7 +869,7 @@ int MarbleMap::northPoleY()
 }
 
 bool MarbleMap::screenCoordinates( qreal lon, qreal lat,
-                                   int& x, int& y )
+                                   qreal& x, qreal& y )
 {
     return d->m_viewParams.currentProjection()
         ->screenCoordinates( lon * DEG2RAD, lat * DEG2RAD,
